requirements.txt
pySerial
pyQt5